change size to return long because of overflow

+++ src/main/groovy/GRange.groovy
@@ -112,7 +112,7 @@ class Region extends Expando implements IRegion, Serializable {
         range instanceof GRange ? range.extra : null
     }
     
-    int size() {
+    long size() {
         range.size()
     }
